requirements.txt
fastapi
uvicorn
pandas==2.2.3
numpy==1.26.4
networkx
pydantic
pytest
python-dotenv
openai
crewai==0.63.6
langchain-openai